cleanup files and directory
